Update to newest Health Data format, refactored to include mapping class, in the mapping sub folder.
