+++ src/main/java/datacollection/mappings/HealthDataFormat.java
@@ -1,7 +1,6 @@
 package datacollection.mappings;
 
 import beans.ActivityMappingBean;
-import datacollection.ActivityMap;
 
 import javax.ejb.EJB;
 import java.text.SimpleDateFormat;
